display initiative, change size -> 32

+++ src/srv/main/content/official/players_handbook/class/rogue.rs
@@ -11,7 +11,7 @@ pub struct Rogue {
 #[content]
 impl Class for Rogue {
     properties! {
-        hit_dice: usize = 8
+        hit_dice: u32 = 8
     }
 
     fn declare(&self, c: &mut Character) {
